backend/requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
gunicorn>=21.0.0
whitenoise>=6.6.0
dj-database-url>=2.1.0
python-decouple>=3.8
psycopg2-binary>=2.9.9
